First version that actually works somewhat.
	modified:   Makefile
	modified:   data.h
	modified:   list_degrees.c
	new file:   main.c
	modified:   pol.c
	modified:   scalar.c
	modified:   test.c
	new file:   xu_and_sparse.c
	new file:   xu_and_sparse.h

+++ data.h
@@ -27,7 +27,7 @@
 
 
 /* This type will be used for our scalars. */
-typedef unsigned int scalar;
+typedef unsigned short int scalar;
 
 struct pol {
 	unsigned int degree;
